Rework handling of "mode-mouse" option
Rename the "copy-mode" value to "no-copy" and make it take effect during
copy mode (if "mode-mouse" is set to "no-copy" then the mouse cannot be
used to make a selection in copy mode).

Also add another value, "scroll-only", which behaves like "no-copy" but
allows the scroll wheel to be used to enter copy mode and to scroll.